# contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: DriveRAG API
  version: "0.2.0"
  description: |
    Agentic Knowledge Graph RAG system API.
    This is the contract that the backend implements and the frontend consumes.
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.dingdong-rag.com/api/v1
    description: Production
paths:
...
  /admin/ingestion/providers:
    get:
      operationId: listIngestionProviders
      summary: List available ingestion providers
      description: Returns all registered source connectors with their configuration status. Admin only.
      tags: [Admin]
      responses:
        "200":
          description: List of available providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProviderInfo"
        "403":
          description: Admin access required
  /admin/ingestion/trigger:
    post:
      operationId: triggerIngestion
      summary: Trigger a new ingestion job
      description: Admin only. Starts ingesting files from Google Drive.
      tags: [Admin]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerIngestionRequest"
      responses:
        "202":
          description: Ingestion job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJobResponse"
        "403":
          description: Admin access required
        "409":
          description: Ingestion already running
  /admin/ingestion/jobs:
    get:
      operationId: listIngestionJobs
      summary: List ingestion jobs
      tags: [Admin]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated list of ingestion jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJobListResponse"
        "403":
          description: Admin access required
  /admin/ingestion/jobs/{job_id}:
    get:
      operationId: getIngestionJob
      summary: Get ingestion job details
      tags: [Admin]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ingestion job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJobResponse"
        "403":
          description: Admin access required
        "404":
          description: Job not found
  /admin/ingestion/jobs/{job_id}/cancel:
    post:
      operationId: cancelIngestionJob
      summary: Cancel a running ingestion job
      tags: [Admin]
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJobResponse"
        "403":
          description: Admin access required
        "404":
          description: Job not found

  # ── Settings ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ProviderInfo:
      type: object
      required: [key, label, configured]
      properties:
        key:
          type: string
          description: Unique source connector key (e.g. 'google_drive').
        label:
          type: string
          description: Human-readable label for the UI.
        configured:
          type: boolean
          description: Whether the required credentials/config are present.
    TriggerIngestionRequest:
      type: object
      properties:
        source:
          type: string
          default: google_drive
          description: Source connector key. Defaults to 'google_drive' for backward compatibility.
        folder_id:
          type: string
          nullable: true
          description: Google Drive folder ID. null = use configured default.
        force:
          type: boolean
          default: false
          description: Re-ingest files even if already processed.
    IngestionJobResponse:
      type: object
      required: [id, source, status, total_files, processed_files, failed_files, skipped_files, started_at]
      properties:
        id:
          type: string
          format: uuid
        source:
          type: string
        status:
          type: string
          enum: [pending, scanning, processing, completed, failed, cancelled]
        folder_id:
          type: string
          nullable: true
        total_files:
          type: integer
        processed_files:
          type: integer
        failed_files:
          type: integer
        skipped_files:
          type: integer
        error_message:
          type: string
          nullable: true
        metadata:
          type: object
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
    IngestionJobListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/IngestionJobResponse"
        total:
          type: integer

    # ── Settings ──────────────────────────────────────────────────────
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...